BFT synch: add block to decision function
